requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.1
mongoengine==0.24.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
